requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.10.0
boto3==1.20.46
botocore==1.23.46
cairocffi==1.0.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cssselect2==0.4.1
cssutils==2.3.0
defusedxml==0.7.1
Django==3.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-inlinecss==0.3.0
django-mail-queue==3.2.4
django-model-utils==4.2.0
django-qr-code==3.0.0
django-storages==1.12.3
django-user-agents==0.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
html5lib==1.1
idna==3.3
importlib-metadata==4.10.1
isort==5.10.1
jmespath==0.10.0
lml==0.1.0
Markdown==3.3.6
maxminddb==2.2.0
multidict==6.0.2
openpyxl==3.0.9
Pillow==9.0.0
psycopg2-binary
pycparser==2.21
pyexcel-io==0.5.20
pyexcel-xls==0.5.9
pyexcel-xlsx==0.5.8
PyJWT==2.3.0
pynliner==0.8.0
pyphen==0.12.0
python-dateutil==2.8.2
python-decouple==3.5
python-magic
python-memcached==1.59
pytz==2021.3
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.0
segno==1.4.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tinycss2==1.1.1
twilio==7.5.1
typing_extensions==4.0.1
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
WeasyPrint==52.5
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.7.2
zipp
